Polybar: write (hacky?) logic to handle multi-monitor setups.
For this, we use sizes for the bars in percentage instead of the
absolute pixel value (which had already been done). However, the
drawback to this approach is that the fonts will still be rendered in
the same size across displays that are of different size/resolution. To
fix this, we remove sizes from the font specification and introduce two
new variables - font-ratio and glyph-ratio. This is the ratio of the
text font to the bar height and the ratio of the glyph font to the bar
font. This is then parsed by the helper script, which will handle sizing
of both text and glyph fonts for all connected/active monitors.

Also, the alternating bar logic has also been updated to support
multi-monitor setups.