environment: introduce new environment package.
The environment package is supposed to be a store for all the central
configuration and also serve as a source of truth for most other
components in this repo.

The structure is dictated by the .profile script, which first sources
the system profile and then looks for the profile directory.
In this directory, the split is made into two, shell profile scripts,
which lives in the profile.d directory and the GUI profile scripts,
which lives in the gui subdirectory within the profile.d directory.

Also there is an exclusion mechanism in the .profile script i.e. any
profile script that needs to be excluded from sourcing can have a
.nosource suffix/extension to the filename; these will be skipped at
profile sourcing and thus can be individually sourced from other places
explicitly instead.